Updated with info about the new desktopdev branch
Branches:
The 'master' version is stable, but is possibly not the latest
version of this project.
The 'dev' branch is the latest version of the web-based version
of this project.
The 'desktopdev' branch is used for developing a version of this
project which is independent of the internet (i.e. it only uses
files on the user's local computer). Being dependent on the
internet is too dangerous in live concert situations.
The 'dev' and 'desktopdev' branches are too volatile to have
public websites.